backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic[email]
httpx
passlib[bcrypt]
pyjwt
python-multipart